project/ai_news/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
google-generativeai
python-dotenv
streamlit
gunicorn
supabase
email-validator
pyjwt

